feat: move SkuDescription to common package
