+++ apps/inventory/src/pages/StockItemDetails.tsx
@@ -14,10 +14,10 @@ import {
 } from '@commercelayer/app-elements'
 import { Link, useLocation, useRoute } from 'wouter'
 
-import { SkuDescription } from '#components/SkuDescription'
 import { StockItemInfo } from '#components/StockItemInfo'
 import { appRoutes } from '#data/routes'
 import { useStockItemDetails } from '#hooks/useStockItemDetails'
+import { SkuDescription } from 'dashboard-apps-common/src/components/SkuDescription'
 import { useState, type FC } from 'react'
 
 export const StockItemDetails: FC = () => {
@@ -128,7 +128,7 @@ export const StockItemDetails: FC = () => {
         <Spacer bottom='4'>
           {stockItem.sku != null && (
             <Spacer top='14'>
-              <SkuDescription sku={stockItem.sku} />
+              <SkuDescription resource={stockItem.sku} />
             </Spacer>
           )}
           <Spacer top='14'>
