+++ apps/price_lists/package.json
@@ -18,6 +18,7 @@
     "@commercelayer/app-elements": "2.9.0",
     "@commercelayer/sdk": "6.22.0",
     "@hookform/resolvers": "^3.9.0",
+    "dashboard-apps-common": "workspace:*",
     "lodash": "^4.17.21",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
