+++ apps/price_lists/src/pages/PriceDetails.tsx
@@ -16,9 +16,9 @@ import { Link, useLocation, useRoute } from 'wouter'
 
 import { PriceInfo } from '#components/PriceInfo'
 import { PriceTiers } from '#components/PriceTiers'
-import { SkuDescription } from '#components/SkuDescription'
 import { appRoutes } from '#data/routes'
 import { usePriceDetails } from '#hooks/usePriceDetails'
+import { SkuDescription } from 'dashboard-apps-common/src/components/SkuDescription'
 import { useState } from 'react'
 
 export function PriceDetails(): JSX.Element {
@@ -126,7 +126,7 @@ export function PriceDetails(): JSX.Element {
         <Spacer bottom='4'>
           {price.sku != null && (
             <Spacer top='14'>
-              <SkuDescription sku={price.sku} />
+              <SkuDescription resource={price.sku} />
             </Spacer>
           )}
           <Spacer top='14'>
