+++ packages/common/package.json
@@ -13,6 +13,7 @@
     "@commercelayer/sdk": "^6.22.0",
     "@hookform/resolvers": "^3.9.0",
     "date-fns": "^4.1.0",
+    "lodash": "^4.17.21",
     "react": "^18.3.1",
     "react-hook-form": "7.53.0",
     "react-qrcode-logo": "^3.0.0",
@@ -22,6 +23,7 @@
   },
   "devDependencies": {
     "@commercelayer/eslint-config-ts-react": "^1.4.5",
+    "@types/lodash": "^4.17.10",
     "@types/react": "^18.3.11",
     "eslint": "^8.57.0",
     "typescript": "^5.6.3"
